Add migration_type and hidden to Migration database model
This adds an enumeration for migration_type to the schema. It does not
do any data translation in the schema migration, and allows the column
to be nullable so that we don't have to act on existing migrations.
This means that any outstanding migrations in the database will need to
be handled lazily by the upper layers.

This also adds a flag called hidden, defaulting to False. Since we are
using this new type field to record potentially admin-initiated actions,
it is desirable to allow them to be hidden from regular users at times.
This adds that field to the database so that we can support that in the
migration APIs.

Related to blueprint robustify-evacuate

Related Upstream Commit: 5c151d08072cdf49094c2056bdce0e37dcaa5274

Signed-off-by: blkart <blkart.org@gmail.com>