compute: rename ResizeClaim to MoveClaim
We will introduce claiming for live-migrate and evacuate operations, and
the claiming logic is really the same for all of those (we will create
different Migration object though, but this is unrelated to claims).

Make sure that the Claim subclass name reflects this fact.

In addition rename a method on the resource tracker for dropping such
claims to 'drop_move_claim' to match.

Related-blueprint: migration-fix-resource-tracking
Related Upstream Commit: d2bbee1718f1bca686f649611ab4d623bd1106ff

Signed-off-by: plysan <plysab@gmail.com>