rt: generalize claim code to be useful for other move actions
Make sure that we can re-use all the code that is needed for doing
claims for live migration and evacuate.

Since we now pass the migration_type to the method that creates it in
the resource tracker, we also re-use some logic that was already in
there to make sure we migration_type is always consistent.

Related-blueprint: migration-fix-resource-tracking
Related Upstream Commit: a414212820fbf60747a94513af391063c8368826

Signed-off-by: plysan <plysab@gmail.com>