rebuild: RPC sends additional args and claims are done
If a migration record was created for an evacuation operation, we will
want to pass it over to the compute node. It will be required for
resource tracking, and we can avoid a round trip back to the conductor
by making it a part of the RPC call.

Also - if we called the scheduler, we also know the node of the chosen
host, and the limits set, and this information is needed for doing claims.

We also make sure that rebuild now uses claims and resource tracking.

There is a final piece of the puzzle missing, and that is the following
patch which actually makes resource tracker update it's resources based
on the evacuation records, and adds the relevant tests.

Related-bug: #1417667
Related-blueprint: migration-fix-resource-tracking
Related Upstream Commit: ad82450059214672c0d294d7cbd2f3ed0213b078
Related Upstream Commit: dc0221d7240326a2d1b467e2a367bebb7e764e61

Signed-off-by: plysan <plysab@gmail.com>