Claims: Make sure move claims create a migration context records
This way we will be able to keep track of the resources that were
actually claimed for those resources (such as NUMA topology) which are
not just a simple sum but actually have state.

This puts in the infra for tracking the claimed resources, however 2
additional steps are needed to fully use this and make resource tracking
work for all instances when migrating:

1) We don't use the resources claimed and stashed in the Migration
context to do resource tracking. This will be a follow-on fix in the
resource tracker.

2) We still don't properly assign the newly claimed topology to the
instance once the migration is confirmed - this will be done in a follow
up patch that builds on this work.

Cherry-pick note: missing test_tracker.py

Partial-bug: #1417667
Related-blueprint: migration-fix-resource-tracking
Related Upstream Commit: c458921729821f3d2aad5a9b76fbc1f80dcd8962

Signed-off-by: plysan <plysab@gmail.com>