Fix instance_extra backref
This fixes cruft from the original NUMA patches that placed a backref on
the Instance model for numa_topology. This changes it to the generic
'instance_extra'. Since this was not actually used, nothing breaks; the
extra backref will be used in subsequent patches.

Related to blueprint flavor-from-sysmeta-to-blob
Related Upstream Commit: 208054a7ef9678dbd4e853f15638e9fad87ea9df

Signed-off-by: plysan <plysab@gmail.com>