Allow strategic loading of InstanceExtra columns
This makes it possible to pull out just one of the several Text
columns with JSON blobs in instance_extra. Before this, we always
pull all the columns even though we only ever want one at a time.
This involves marking the big columns in InstanceExtra as deferred(),
and specifying which ones we want each time. This is hidden in the
API for querying, but should be noted in any joins.

Related to blueprint flavor-from-sysmeta-to-blob
Related Upstream Commit: 8cf9dfe3fd759304beccc92b82310d61040925ec

Signed-off-by: plysan <plysab@gmail.com>