RT: track evacuation migrations
The previous patch sets the stage for tracking resources when doing
evacuate, but we still have to make sure that
_update_usage_from_migration actually does the update of the resources.
To do that we have to make sure we are getting the right flavor in case
of 'evacuation' migration, and also that we are not completely skipping
over it.

Once that's done claims and resource tracking work as expected when
evacuating instances.

Related Upstream Commit: 4ee4f9f2ec639d3361d65dd2187f5bde48be9970
Related-bug: #1417667

Signed-off-by: blkart <blkart.org@gmail.com>